+++ packages/app-bridge/src/AppBridgePlatformApp.spec.ts
@@ -11,8 +11,10 @@ const TOKEN = 'AjY34F87Dsat^J';
 describe('AppBridgePlatformApp', () => {
     vi.mock('./utilities/subscribe', () => ({
         subscribe: vi.fn().mockResolvedValue({
-            port: { onmessage: vi.fn() },
+            statePort: { onmessage: vi.fn() },
+            apiPort: { onmessage: vi.fn() },
             context: { parentId: 'parentId-test', connected: true },
+            state: { settings: 'settings-test' },
         }),
     }));
 
@@ -37,14 +39,6 @@ describe('AppBridgePlatformApp', () => {
         expect(notify).toHaveBeenCalledTimes(0);
     });
 
-    it('should notify endpoint when openConnection is called correctly', async () => {
-        window.location.search = `?token=${TOKEN}`;
-
-        const platformApp = new AppBridgePlatformApp();
-        await platformApp.dispatch(openConnection());
-        expect(notify).toHaveBeenCalledTimes(1);
-    });
-
     it('should yield true for Context.connected after dispatch', async () => {
         const connected = true;
         window.location.search = `?token=${TOKEN}`;
@@ -62,10 +56,22 @@ describe('AppBridgePlatformApp', () => {
         await expect(() => platformApp.api({ name: 'getCurrentUser' })).rejects.toThrow();
     });
 
-    it('should return undefined state as it is not implemented', async () => {
+    it('should return empty state when not inititalized', async () => {
         const platformApp = new AppBridgePlatformApp();
-        const state = platformApp.state();
-        expect(state).toEqual(undefined);
+        const state = platformApp.state().get();
+        expect(state).toEqual({ settings: {} });
+    });
+
+    it('should return state after app is initialized', async () => {
+        window.location.search = `?token=${TOKEN}`;
+        const platformApp = new AppBridgePlatformApp();
+        platformApp.subscribe('Context.connected', (connected) => {
+            const state = platformApp.state().get();
+
+            expect(connected).toBe(true);
+            expect(state).toEqual({ settings: 'settings-test' });
+        });
+        platformApp.dispatch(openConnection());
     });
 
     it('should yield true for Context.connected after dispatch', async () => {
@@ -87,6 +93,16 @@ describe('AppBridgePlatformApp', () => {
         const platformApp = new AppBridgePlatformApp();
         platformApp.context().subscribe((context) => {
             expect({ parentId: 'parentId-test', connected: true }).toStrictEqual(context);
+        });
+        platformApp.dispatch(openConnection());
+    });
+
+    it('should return correct object when subscribing to context', async () => {
+        window.location.search = `?token=${TOKEN}`;
+        const platformApp = new AppBridgePlatformApp();
+        platformApp.context().subscribe((context) => {
+            expect({ parentId: 'parentId-test', connected: true }).toStrictEqual(context);
+            platformApp.dispatch(openConnection());
         });
         platformApp.dispatch(openConnection());
     });