+++ packages/guideline-blocks-settings/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @frontify/guideline-blocks-settings
 
+## 0.29.13
+
+### Patch Changes
+
+- Updated dependencies [[`dfe83ee`](https://github.com/Frontify/brand-sdk/commit/dfe83ee7417d3916d9bc52b96e600deb130bf147), [`8920ec9`](https://github.com/Frontify/brand-sdk/commit/8920ec993401caf94813974c7fad476c3decc2c6)]:
+  - @frontify/sidebar-settings@0.7.0
+
+## 0.29.12
+
+### Patch Changes
+
+- [#628](https://github.com/Frontify/brand-sdk/pull/628) [`44efda9`](https://github.com/Frontify/brand-sdk/commit/44efda90fd454298ea623c7f41b370f81ad1a203) Thanks [@fulopdaniel](https://github.com/fulopdaniel)! - fix(BlockItemWRapper): fix toolbar focus style
+
+## 0.29.11
+
+### Patch Changes
+
+- [#622](https://github.com/Frontify/brand-sdk/pull/622) [`34aa320`](https://github.com/Frontify/brand-sdk/commit/34aa32063c306826d35a6b973c3e5f6931dabc76) Thanks [@fulopdaniel](https://github.com/fulopdaniel)! - fix(BlockItemWrapper): add focus visible style to drag btn
+
 ## 0.29.10
 
 ### Patch Changes